% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{asinh_trans}
\alias{asinh_trans}
\title{asinh ggplot scale transformation}
\usage{
asinh_trans()
}
\description{
This can be used to do an arcsinh transformation of a ggplot scale, similar to the log transformation (but with the ability to handle values <= 0). To do a sinh transformation, use "sinh" instead of "asinh" (see ?sinh_trans).
}
\examples{
library(ggplot2)
ggplot(data.frame(x=1:10,y=1:10), aes(x=x,y=y)) +
    geom_point() +
    scale_y_continuous(trans="asinh")

}
