% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_funcs.R
\name{bin_to_raster}
\alias{bin_to_raster}
\title{Generate 2D grid of L3b data for viewing}
\usage{
bin_to_raster(
  df,
  resolution = "4",
  ext = c(xmn = -147, xmx = -41, ymn = 39, ymx = 86),
  rast = TRUE,
  max_bins = 5e+07
)
}
\arguments{
\item{df}{Dataframe with 2 columns: first column="bin" and second column is the variable name (note: the dataframe does not have to be sorted in order of bin number).}

\item{resolution}{String indicating spatial resolution, see ?gen_nrows for list of accepted strings.}

\item{ext}{Named vector containing the boundaries of the resulting grid (xmn, xmx, ymn, ymx).}

\item{rast}{TRUE/FALSE, should the resulting bin matrix be converted to raster?}

\item{max_bins}{Maximum grid size (total number of pixels) that can be generated between the selected latitudes. This is to prevent overloading memory with the higher-resolution grids (e.g. 250m, 1km)}
}
\value{
Raster containing the variable data.
}
\description{
Given a dataframe with 2 columns (bin number and variable), map the results to an equidistant cylindrical grid using a nearest-neighbour method.
}
\details{
Originally coded by George White.
}
\examples{
# make a bathymetry raster for the Northwest Atlantic
bin_to_raster(df = get_bins(region="nwa",variables=c("bin","bathymetry")),
              ext = c(lon_bounds$NWA, lat_bounds$NWA))
}
\references{
See https://oceancolor.gsfc.nasa.gov/docs/format/l3bins/ for more information on bin numbers.
}
