% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{calculate_PAAW}
\alias{calculate_PAAW}
\title{Calculate PAAW (Phytoplankton Apparent Absorption Wavelength)}
\usage{
calculate_PAAW(abs)
}
\arguments{
\item{abs}{Numeric vector, length 301: phytoplankton absorption (1/m) corresponding to wavelengths 400 to 700 nm}
}
\value{
Single numeric value
}
\description{
Given a numeric vector of phytoplankton absorption values from 400 to 700 nm at 1 nm intervals, calculate the PAAW metric as in Devred et al. (2022).
}
\examples{
\dontrun{
abs <- c(0.045,0.046,0.046,0.047,0.047,0.047,0.048,0.048,0.048,0.049,0.049,0.049,0.05,0.05,0.05,0.05,0.051,0.051,0.051,0.051,0.051,0.051,0.051,0.051,0.051,0.051,0.051,0.051,0.052,0.052,0.052,0.053,0.053,0.054,0.054,0.055,0.055,0.055,0.055,0.055,0.055,0.054,0.053,0.053,0.052,0.051,0.05,0.049,0.049,0.048,0.047,0.047,0.046,0.046,0.045,0.045,0.044,0.044,0.044,0.043,0.043,0.043,0.042,0.042,0.042,0.041,0.041,0.04,0.04,0.039,0.039,0.038,0.037,0.037,0.036,0.035,0.035,0.034,0.033,0.033,0.032,0.032,0.032,0.031,0.031,0.031,0.031,0.03,0.03,0.03,0.029,0.029,0.029,0.028,0.028,0.028,0.028,0.028,0.027,0.027,0.027,0.027,0.026,0.026,0.025,0.025,0.025,0.024,0.024,0.023,0.023,0.022,0.022,0.022,0.021,0.021,0.021,0.02,0.02,0.02,0.02,0.019,0.019,0.019,0.018,0.018,0.018,0.017,0.017,0.017,0.017,0.017,0.017,0.016,0.016,0.016,0.016,0.015,0.015,0.015,0.015,0.014,0.014,0.014,0.013,0.013,0.013,0.013,0.012,0.012,0.012,0.011,0.011,0.011,0.01,0.01,0.01,0.01,0.009,0.009,0.009,0.009,0.008,0.008,0.008,0.008,0.008,0.008,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.007,0.006,0.006,0.006,0.006,0.006,0.006,0.007,0.007,0.007,0.007,0.007,0.008,0.008,0.008,0.008,0.008,0.009,0.009,0.009,0.009,0.009,0.009,0.009,0.009,0.009,0.009,0.009,0.01,0.01,0.01,0.01,0.01,0.01,0.01,0.01,0.01,0.01,0.01,0.01,0.01,0.009,0.009,0.009,0.009,0.009,0.009,0.009,0.009,0.009,0.009,0.009,0.009,0.009,0.01,0.01,0.011,0.011,0.012,0.013,0.015,0.016,0.017,0.019,0.021,0.022,0.024,0.025,0.027,0.028,0.029,0.03,0.031,0.032,0.032,0.033,0.033,0.032,0.032,0.031,0.03,0.029,0.027,0.026,0.024,0.023,0.021,0.019,0.017,0.015,0.013,0.011,0.01,0.008,0.007,0.006,0.004,0.003,0.003,0.002,0.001,0.001,0)
calculate_PAAW(abs)
}
}
\references{
Devred, E., Perry, T., and Massicotte, P., "Seasonal and decadal variations in absorption properties of phytoplankton and non-algal particulate matter in three oceanic regimes of the Northwest Atlantic", Frontiers in Marine Science, vol. 9, Art. no. 932184, 2022. doi:10.3389/fmars.2022.932184.
}
