% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hu.R
\name{conv_rrs_to_555}
\alias{conv_rrs_to_555}
\title{Convert green Rrs to Rrs at 555nm}
\usage{
conv_rrs_to_555(rrs, wave)
}
\arguments{
\item{rrs}{Numeric vector of remote sensing reflectances (Rrs) from the "green" waveband}

\item{wave}{Numeric value, "green" waveband associated with these rrs values (example: 547)}
}
\value{
Numeric vector of Rrs converted to 555nm
}
\description{
Given a numeric vector of Rrs at a green waveband not centered at 555nm, convert the Rrs to Rrs at 555nm.
}
\references{
Adapted from function coded at NASA OBPG. Source:

https://oceancolor.sci.gsfc.nasa.gov/docs/ocssw/convert__band_8c_source.html
}
