% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_funcs.R
\name{convert_daily_grid}
\alias{convert_daily_grid}
\title{Group daily bin matrix as 8day or monthly}
\usage{
convert_daily_grid(mat, composite = "8day")
}
\arguments{
\item{mat}{Numeric matrix where rows=binned pixels and columns=days of the year (in order, with no gaps - if a day has no data, it should be a column of all NA)}

\item{composite}{String, length of output composite, 8day or monthly}
}
\value{
Numeric matrix where rows = binned pixels, columns = weeks (8day) or months
}
\description{
Input a matrix where rows = binned pixels and columns = days of the year, and use rowMeans to condense the columns to 8day or monthly composites, resulting in a matrix of (binned pixels) x (months), for example. Note that a day that has no valid data should be a column of all NA, but if consecutive days are missing from the end of the year (for example, if the data hasn't been made available yet), those columns could be excluded. Example: If the dataset for 2020 ends at day 321 (Nov. 16th), and you're using the panCanadian grid which has 529797 pixels, the matrix should have dimensions 529797 x 321.
}
