% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_blank_raster.R
\name{create_blank_raster}
\alias{create_blank_raster}
\title{Blank raster}
\usage{
create_blank_raster(ROI)
}
\arguments{
\item{ROI}{String, region of interest (either "AZMP" or "extNA").}
}
\value{
Blank raster bound to AZMP template.
}
\description{
Create a basic template for a region of interest (ROI, currently only AZMP or extNA).
}
