% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_funcs.R
\name{dategrp}
\alias{dategrp}
\title{Date group numbers}
\usage{
dategrp(dateval, grp_length)
}
\arguments{
\item{dateval}{Date string or date formatted with as.Date()}

\item{grp_length}{Number indicating length of group of days}
}
\value{
numeric value of "week" number
}
\description{
This assigns a date or date vector to a corresponding group of dates of some length in a given year. This is a helper function for making composites. The group is calculated from day 1 of the year.
If no group length entered, this just uses the lubridate::week() function (7-day week length)
}
\examples{
dategrp(dateval = "2020-03-27")
dategrp(dateval = c("2020-03-27","2020-04-27"), grp_length = 4)
}
