% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_funcs.R
\name{days_vector}
\alias{days_vector}
\title{List days of the year within a month}
\usage{
days_vector(year, month)
}
\arguments{
\item{year}{Numeric 4-digit value}

\item{month}{Numeric value (1-12)}
}
\value{
Vector of numeric values (days of the year)
}
\description{
Get a list of days of the year within a specific month
}
