% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{deg_to_dm}
\alias{deg_to_dm}
\title{Convert a latitude or longitude from decimal degrees to degrees, decimal minutes}
\usage{
deg_to_dm(decdeg)
}
\arguments{
\item{decdeg}{number in decimal degrees}
}
\value{
Returns a data.frame() with columns indicating degrees and decimal minutes. For degrees, minutes, seconds use \code{deg_to_dms()}
}
\description{
Convert a latitude or longitude from decimal degrees to degrees, decimal minutes
}
\examples{
deg_to_dm(-52.80833)

}
