% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{deg_to_dms}
\alias{deg_to_dms}
\title{Convert a latitude or longitude from decimal degrees to degrees, minutes, seconds}
\usage{
deg_to_dms(decdeg)
}
\arguments{
\item{decdeg}{number in decimal degrees}
}
\value{
Returns a data.frame() with columns indicating degrees, minutes, seconds. For degrees and decimal minutes use \code{deg_to_dm()}
}
\description{
Convert a latitude or longitude from decimal degrees to degrees, minutes, seconds
}
\examples{
deg_to_dms(-52.80833)

}
