% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_funcs.R
\name{dms_to_deg}
\alias{dms_to_deg}
\title{Convert a latitude or longitude in degrees, minutes, seconds, to decimal degrees}
\usage{
dms_to_deg(deg, min, sec)
}
\arguments{
\item{deg}{degrees}

\item{min}{minutes}

\item{sec}{seconds}
}
\value{
Input the degrees, minutes and seconds to return the decimal degrees.
}
\description{
Also works for degrees, decimal minutes to decimal degrees (leave seconds blank)
}
\details{
To reverse this, use \code{deg_to_dms()} or \code{deg_to_dm()}.
}
\examples{
dms_to_deg(deg = -52, min = 48, sec = 30)
dms_to_deg(-52, 48.5)

}
