% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satfile_funcs.R
\name{download_oc}
\alias{download_oc}
\title{Download ocean colour files}
\usage{
download_oc(filenames, urls)
}
\arguments{
\item{filenames}{Filename to download}

\item{urls}{URLs of files corresponding to filenames}
}
\value{
Data.frame with metadata and URLs of files to download.
}
\description{
Download file lists from functions get_imglist_multi() and get_imglist_l1()
}
\details{
Uses wget (works on Mac, Linux, or Cygwin from Windows), and you need to have the OBPG .netrc credentials, i.e. you need to create an Earthdata account, and then type the following on the command line, replacing USERNAME and PASSWD with your own username and password:

echo "machine urs.earthdata.nasa.gov login USERNAME password PASSWD" > ~/.netrc ; > ~/.urs_cookies

chmod 0600 ~/.netrc

See "Download Methods" here for more details: https://oceancolor.gsfc.nasa.gov/data/download_methods/
}
\examples{
filename = "A2003001162000.L2_LAC_OC.nc"
url = "https://oceandata.sci.gsfc.nasa.gov/cmr/getfile/A2003001162000.L2_LAC_OC.nc"
download_oc(filename, url)
}
