% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{filtered_mean}
\alias{filtered_mean}
\title{Calculate filtered mean, stdev, and CV}
\usage{
filtered_mean(var)
}
\arguments{
\item{var}{Numeric vector or matrix}
}
\value{
Named list containing the filtered mean, filtered standard deviation, and coefficient of variation
}
\description{
Given a numeric vector, calculate the filtered mean, filtered standard deviation, and coefficient of variation (filtered standard deviation over filtered mean)
}
\examples{
filtered_mean(rnorm(50))
filtered_mean(matrix(rnorm(60),nrow=12,ncol=5))

}
