% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{find_factors}
\alias{find_factors}
\title{Find all the factors of a positive integer}
\usage{
find_factors(x)
}
\arguments{
\item{x}{Positive integer}
}
\value{
Vector of integers containing all the positive factors of x (from 1 to x inclusive)
}
\description{
Find all the factors of a positive integer
}
\examples{
find_factors(112749)

}
