% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{find_line}
\alias{find_line}
\title{Compute line parameters}
\usage{
find_line(x1, y1, x2, y2)
}
\value{
Named list containing slope and intercept.
}
\description{
Find slope and intercept of a line based on the coordinates of two points.
}
