% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_funcs.R
\name{gen_nrows}
\alias{gen_nrows}
\title{Get number of rows on L3b global grid for a given resolution}
\usage{
gen_nrows(resolution)
}
\arguments{
\item{resolution}{String representing the spatial resolution (see details for list of accepted strings, e.g. "Q" for 250m or "4" for 4.6km)}
}
\value{
Single numeric value, the number of rows
}
\description{
This generates the number of rows on the global grid of level-3 binned (L3b) data for a given spatial resolution. For example: For 4km resolution data, 4320 rows; for 9km resolution, 2160 rows
}
\details{
HH: 50m (currently unavailable)
HQ: 100m (currently unavailable)
Q: 250m
H: 0.5km
1: 1.1km
2: 2.3km
4: 4.6km
9: 9.2km
18: 18.5km
36: 36km
QD: 0.25 degree (currently unavailable)
HD: 0.5 degree
1D: 1 degree
}
\references{
https://oceancolor.gsfc.nasa.gov/docs/format/l3bins/
https://oceancolor.gsfc.nasa.gov/docs/ocssw/
}
