% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{geoMean}
\alias{geoMean}
\title{Geometric mean}
\usage{
geoMean(x, ...)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Geometric mean value of the input data
}
\description{
This calculates the geometric mean, AKA mean of the log transformed data converted back to given units. Use with lognormal distributions like chlorophyll-a
}
