% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHLA_GSM.R
\name{get_bbw}
\alias{get_bbw}
\title{Get water backscattering coefficients (bbw) for selected wavebands}
\usage{
get_bbw(lambda)
}
\arguments{
\item{lambda}{Numeric vector of wavebands (nanometers)}
}
\value{
Numeric vector of bbw coefficients corresponding to lambda
}
\description{
If lambda contains wavebands that are not integer values, the value will be interpolated. Bands must be within 400-700nm (inclusive).
}
\references{
Sources for default aw, bbw, and aphstar: Pope and Fry 1997, and Smith and Baker 1981 (https://oceancolor.gsfc.nasa.gov/docs/rsr/water_coef.txt, this does not account for salinity effects on backscattering like the values in Zhang 2009), and DFO cruise records containing aph and chlorophyll-a values which were converted to aphstar by mean(aph/chl).
}
