% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_funcs.R
\name{get_bins}
\alias{get_bins}
\title{Get bin info at 4km and 9km resolution}
\usage{
get_bins(region = "pancan", resolution = "4km", variables = "all")
}
\arguments{
\item{region}{String, either "pancan", "nwa", "nep", or "gosl"}

\item{resolution}{String, either "4km" or "9km".}

\item{variables}{String with columns: blank or "all" for all columns, or a subset of c("bin","longitude","latitude","bathymetry")}
}
\value{
Data frame with columns of bin, longitude, latitude, bathymetry
}
\description{
Get corresponding bin number, latitude, longitude and depth for Pan-Canadian Grid or subregions. These are vectors of bin/lat/lon/bathymetry that have already been generated for easier access.
}
\examples{
library(dplyr)
library(ggplot2)
library(oceancolouR)

pancan_bins_4km <- get_bins()

# Map of the North West Atlantic bin bathymetry
get_bins(region = "nwa", resolution = "9km") \%>\%
ggplot(aes(x = longitude, y = latitude, colour = bathymetry)) +
   geom_point(size = 0.5) +
   scale_colour_gradientn(colours = pals::ocean.deep(30))
}
