% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCX.R
\name{get_br}
\alias{get_br}
\title{Get band ratios}
\usage{
get_br(rrs, blues, green, use_443nm = FALSE)
}
\arguments{
\item{rrs}{Numeric matrix where rows = records, columns = Rrs wavebands, with named columns ("Rrs_XXX", where XXX is a wavelength in nanometres)}

\item{blues}{Character vector of Rrs wavebands in the blue range (e.g. "Rrs_488"), matching column name(s) in rrs}

\item{green}{Character vector of Rrs waveband in the green range (e.g. "Rrs_547"), matching a column name in rrs}

\item{use_443nm}{Logical value, TRUE to make the 443nm band an option in the band ratio}
}
\value{
Named list of three vectors: rrs_ocx (the band ratio values), ind (the index of valid rrs values), and ratio_used (strings indicating which "blue" waveband was used for each band ratio)
}
\description{
Given a matrix of Rrs (remote sensing reflectances) with column names, and the blue and green wavebands to use, calculate the band ratios.
}
