% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHLA_Hu.R
\name{get_ci_bands}
\alias{get_ci_bands}
\title{Get wavebands to use in the color index algorithm}
\usage{
get_ci_bands(sensor)
}
\arguments{
\item{sensor}{String, either "modis" (for MODIS-Aqua), "seawifs", or "viirs" (for VIIRS-SNPP)}
}
\value{
Numeric vector of wavebands for the selected sensor, to use in the color index and Hu algorithm.
}
\description{
Given a sensor name, get the appropriate wavebands closest to 443, 555, and 670nm. Note that for MODIS, the 547nm band should be used for the green band (547nm is the "ocean" band, 555nm is the "land" band).
}
\references{
Link to NASA chlorophyll-a description:

https://oceancolor.gsfc.nasa.gov/atbd/chlor_a/
}
