% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHLA_Hu.R
\name{get_ci_coefs}
\alias{get_ci_coefs}
\title{Get coefficients to use in the Hu chlorophyll-a algorithm.}
\usage{
get_ci_coefs(version)
}
\arguments{
\item{version}{Single numeric value, either 1 or 2}
}
\value{
Numeric vector of coefficients, from lowest degree to highest.
}
\description{
Given a version number (either 1 or 2), get the Hu coefficients. Version 1 coefficients are from the 2012 paper, version 2 are from the 2019 paper.
}
\references{
#' Original paper:

Hu, Chuanmin & Lee, Zhongping & Franz, Bryan. (2012). Chlorophyll a algorithms for oligotrophic oceans: A novel approach based on three-band reflectance difference. Journal of Geophysical Research. 117. C01011. 10.1029/2011JC007395.

Updates:

Hu, C., Feng, L., Lee, Z., Franz, B. A., Bailey, S. W., Werdell, P. J., & Proctor, C. W. (2019). Improving satellite global chlorophyll a data products through algorithm refinement and data recovery. Journal of Geophysical Research: Oceans, 124, 1524– 1543. https://doi.org/10.1029/2019JC014941

Link to NASA chlorophyll-a description:

https://oceancolor.gsfc.nasa.gov/atbd/chlor_a/
}
