% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchup_funcs.R
\name{get_closest_bins}
\alias{get_closest_bins}
\title{Given lat/lon, find closest bin(s)}
\usage{
get_closest_bins(
  geo_df,
  bin_df,
  measure = "geodesic",
  max_bins = 100,
  radius = Inf
)
}
\arguments{
\item{geo_df}{Dataframe with numeric columns longitude, latitude}

\item{bin_df}{Dataframe with numeric columns bin, longitude, latitude (created using the bin, lon, lat vectors in this package, see details)}

\item{measure}{Algorithm to use for distance calculation, see "measure" in ?geodist::geodist}

\item{max_bins}{Maximum number of closest bins to return (<= 100)}

\item{radius}{Maximum distance between lat/lon pair and bin lat/lon (in metres)}
}
\value{
geo_df dataframe with new columns "dist" (distance to closest bin(s), in metres), "bin" (closest bin number(s)), and bin_latitude/longitude
}
\description{
If you have a dataframe of longitudes and latitudes, find the closest bin numbers from NASA's 4km or 9km-resolution L3b (level-3 binned) files.
}
\details{
Bin/lat/lon data can be retrieved using the get_bins() function (see examples).
}
