% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{get_dir}
\alias{get_dir}
\title{Create folder.}
\usage{
get_dir(path)
}
\arguments{
\item{path}{Path to create (a string). Absolute or relative.}
}
\value{
The path name.
}
\description{
Check for existence of a folder, and create it if necessary.
}
