% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_funcs.R
\name{get_doys}
\alias{get_doys}
\title{Get start days, given interval}
\usage{
get_doys(year, months = 1:12, eightday = TRUE, interval = "daily")
}
\arguments{
\item{year}{Numeric value}

\item{months}{Numeric vector of month numbers}

\item{eightday}{Logical value (use eightday system?)}

\item{interval}{String, either "daily", "weekly", "biweekly", or "monthly"}
}
\value{
Character vector containing days of the year, each zero-padded to length 3.
}
\description{
Get a list of days of the year for a given year and selected months. This can return every day, or the start of every week, second week, or month. It also can use either the 8-day/week 46 week/year system as used by NASA, or the 4 week/month 48 week/year system as used at BIO (only relevant for weekly or biweekly intervals).
}
