% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satfile_funcs.R
\name{get_edges}
\alias{get_edges}
\title{Extract the edges of a matrix}
\usage{
get_edges(mat)
}
\arguments{
\item{mat}{Matrix}
}
\value{
Vector of edge values
}
\description{
Get the edge values of a matrix, starting at the top left and moving clockwise.
}
