% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHLA_GSM.R
\name{get_gs}
\alias{get_gs}
\title{GSM gs coefficients}
\usage{
get_gs(lambda)
}
\arguments{
\item{lambda}{Numeric vector of wavelengths in nanometres}
}
\value{
Named list containing the numeric vectors g1, g2, g3 corresponding to input lambda
}
\description{
Given a numeric vector of wavelengths, calculate the corresponding g1, g2, and g3 coefficients to use in the GSM model.
}
