% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSM.R
\name{get_gsm_IOPexps}
\alias{get_gsm_IOPexps}
\title{Get predefined IOP exponents for GSM}
\usage{
get_gsm_IOPexps(sensor, region, gtype)
}
\arguments{
\item{sensor}{String, either "modis", "seawifs", or "viirs" (note: "modis" is MODIS-Aqua, and "viirs" is VIIRS-SNPP)}

\item{region}{String, either "global", "nwa", or "nep"}

\item{gtype}{String, either "gc" or "gs" (see description below)}
}
\value{
Numeric vector of exponents used on the IOPs in the algorithm, in this order: chl, adg, bbp
}
\description{
Library of existing optimized exponents for the Garver-Siegel-Maritorena algorithm, used on the IOPs chl, adg, and bbp within the absorption and backscattering terms of the algorithm.
}
\details{
region="global" for the standard exponents, "nwa" (Northwest Atlantic) or "nep" (Northeast Pacific) for the regionally-tuned exponents used in Clay et al 2019.

gtype="gc" for the exponents that were tuned using the g coefficients that are constant across wavebands, "gs" for those tuned using the spectrally-dependent g coefficients described in Clay et al 2019.
}
\references{
Clay, S.; Peña, A.; DeTracey, B.; Devred, E. Evaluation of Satellite-Based Algorithms to Retrieve Chlorophyll-a Concentration in the Canadian Atlantic and Pacific Oceans. Remote Sens. 2019, 11, 2609.
https://www.mdpi.com/2072-4292/11/22/2609
}
