% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satfile_funcs.R
\name{get_imglist_l1}
\alias{get_imglist_l1}
\title{Get L1A filenames}
\usage{
get_imglist_l1(
  sensor = "MODISA",
  minlat,
  maxlat,
  minlon,
  maxlon,
  mindate,
  maxdate
)
}
\arguments{
\item{sensor}{Currently works for options "MODISA" and "SEAWIFS"}

\item{minlat}{Minimum latitude (y)}

\item{maxlat}{Maximum latitude (y)}

\item{minlon}{Minimum longitude (x)}

\item{maxlon}{Maximum longitude (x)}

\item{mindate}{Earliest image as string of year or date (YYYY-MM-dd)}

\item{maxdate}{Latest image as string of year or date (YYYY-MM-dd)}
}
\value{
Data.frame with metadata and URLs of files to download.
}
\description{
Query the CMR Search site for files falling in a specific area. This currently converts the L2 names to L1 as the L1 files aren't fully listed on the CMR site.
}
