% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satfile_funcs.R
\name{get_imglist_multi}
\alias{get_imglist_multi}
\title{Get L2 / L3 filenames}
\usage{
get_imglist_multi(dataset, minlat, maxlat, minlon, maxlon, mindate, maxdate)
}
\arguments{
\item{dataset}{Code indicating what dataset: see https://cmr.earthdata.nasa.gov/search/site/collections/directory/OB_DAAC/gov.nasa.eosdis for full list}

\item{minlat}{Minimum latitude (y)}

\item{maxlat}{Maximum latitude (y)}

\item{minlon}{Minimum longitude (x)}

\item{maxlon}{Maximum longitude (x)}

\item{mindate}{Earliest image as string of year or date (YYYY-MM-dd)}

\item{maxdate}{Latest image as string of year or date (YYYY-MM-dd)}
}
\value{
Data.frame with metadata and URLs of files to download.
}
\description{
Query the CMR Search site for files falling in a specific area. Most codes should work. A few useful ones:
\itemize{
\item MODISA_L2_OC (_IOP, _SST)
\item OLCIS3A_L2_EFR_OC (S3A_) (\emph{ERR}) (_IOP, _OC_NRT, _IOP_NRT)
\item VIIRSN_L2_OC (VIIRSJ1_) (_IOP, _SST)
\item SEAWIFS_L2_OC (_IOP)
\item CZCS_L2_OC
\item MODISA_L3b_SST (_SST_NRT)
\item MODISA_L3b_NSST (_NSST_NRT)
\item MODISA_L3b_SST4 (_SST4_NRT)
\item MODISA_L3b_CHL (_CHL_NRT)
}
}
