% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCX.R
\name{get_lambda}
\alias{get_lambda}
\title{Get OCX wavebands for a sensor}
\usage{
get_lambda(sensor, use_443nm)
}
\arguments{
\item{sensor}{String, either "modis", "seawifs", or "viirs"}

\item{use_443nm}{Logical value, TRUE to make the 443nm band an option in the band ratio}
}
\value{
Named list of 2 character vectors, one for "green" waveband(s) and one for "blue"
}
\description{
Given a sensor name, get the wavebands typically used in the OCX algorithm.
}
