% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{get_lm_stats}
\alias{get_lm_stats}
\title{Create a dataframe of statistics for a model object}
\usage{
get_lm_stats(lm_obj)
}
\arguments{
\item{modelobject}{Object of class lm}
}
\value{
Dataframe of statistics
}
\description{
Given a linear model object created with lm(), retrieve a dataframe with the
coefficients, Rsquared, p-value, and number of observations.
}
