% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_lon}
\alias{get_lon}
\alias{get_lat}
\alias{get_epoch}
\alias{get_time}
\alias{get_res}
\title{Retrieve coordinates}
\usage{
get_lon(x)

get_lat(x)

get_epoch(x)

get_time(x)

get_res(x)
}
\arguments{
\item{x}{ncdf4 object}
}
\value{
numeric vector
}
\description{
The functions \code{get_lon} and \code{get_lat} return numeric vectors of pixel center locations.
\code{get_epoch} and \code{get_time} return (most likely) Date vectors, but support for
POSIXct in the future is possible. The function \code{get_res} returns a two element
numeric vector of resolution in x and y.
}
