% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{get_ranges}
\alias{get_ranges}
\title{Get ggplot scale ranges}
\usage{
get_ranges(p)
}
\arguments{
\item{p}{A ggplot object}
}
\value{
Named list with x and y axis ranges
}
\description{
Given a ggplot object, get the ranges of the x and y scales.
}
\examples{
library(ggplot2)
p <- ggplot(data.frame(x1=1:100,y1=50:-49), aes(x=x1,y=y1)) + geom_point()
get_ranges(p)

}
