% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_funcs.R
\name{get_season}
\alias{get_season}
\title{Get season name}
\usage{
get_season(date, version = 1)
}
\arguments{
\item{date}{Date object}

\item{version}{Number (1, 2, or 3) - system to use to separate seasons (see Details)}
}
\value{
String (either Spring, Summer, Fall, or Winter)
}
\description{
Given a date object, get the name of the season.
}
\details{
There are 3 different options for separating seasons:
Version 1: Astronomical system (e.g. Spring = March 20th - June 20th)
Version 2: Each season is 3 months, Spring is March-May
Version 3: Each season is 3 months, Spring is Feb-April
}
