% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPM.R
\name{get_spm_nechad}
\alias{get_spm_nechad}
\title{Calculate Nechad SPM}
\usage{
get_spm_nechad(spm_rrs, sensor = "occci")
}
\arguments{
\item{spm_rrs}{Either: Numeric vector of Rrs from the red band used in the model (see ?get_spm_nechad_band), OR: RasterLayer or SpatRaster of these Rrs}

\item{sensor}{String of the sensor name, currently only occci is accepted}
}
\value{
Numeric vector (or RasterLayer or SpatRaster) of the spm values
}
\description{
Given a sensor name, calculate SPM using Nechad's model. Note that the input to this function is Rrs, which is converted to rho within the function by multiplying by pi.
}
\examples{
# example using a vector
spm_rrs <- c(0.00032,0.00044,0.00075,2e-04,0.00023,8e-05,0.00022,5e-05,4e-05,4e-05,1e-05,0.00013,0.00029,0.00043,6e-04,0.00053,0.00048,0.00097,0.0022,0.00238,0.00126,0.00048,0.00171,0.00118,0.00214,0.00027,0.00011,0.00027,0.00029,0.00094,0.00018,0.00134,0.00076,0.00124,0.00083,4e-04,0.00051,0.00027,0.00024,0.00137,0.00079,0.00039,0.00056,0.00167,0.00046,5e-04,0.00056,0.00034,3e-04,0.00039,0.00058,0.00084,0.00042,0.00067,0.00115,0.00093,0.00054,0.00043,0.00061,0.00076,0.00044,0.0013,0.00057,0.00101,0.00106,0.00084,0.00133,0.00103,6e-04,0.00034,0.00017,0.00014,0.00018,0.00029,0.00039,0.00038,0.00048,0.00057,0.00067,0.00048,0.00014,0.00036,0.00018,0.00141,8e-05,0.00141,0.00025,0.00108,0.00045,0.00079,0.00198,0.00115,0.00036,0.00042,0.00025,0.00013,0.00043,3e-04,0.00151,0.00081,0.00116,0.00155,0.00164,8e-05,0.00029,0.00015,0.00078,0.0014,0.00105,0.00103,0.00102,0.00084,2e-04,0.00031,0.00031,0.00045,4e-04,0.00089,0.00078,0.00072,0.00139,0.00034,0.00037,0.00052,0.00046,0.00065,0.00015,3e-04,0.00043,0.00013,0.00058,0.00016,0.00021,0.00022,0.00061,0.00045,0.0014,0.0035,0.00239,0.00694,0.00142,8e-05,0.00036,0.00035,0.00069,0.00043,0.00049,0.00046,0.00245,0.00074,0.00107,0.00087,0.00025,0.00027,0.00022,0.00022,0.00027,0.00059,0.00069,0.00033,0.00055,0.00049,0.00034,0.00028,0.00131,0.00202,0.00074,0.00083,0.00043,0.00042,0.00101,9e-04,0.00111,0.00094,0.00031,0.00025,0.00027,0.00024,2e-04,0.00022,0.00055,0.00013,0.00137,0.00115,0.00152,0.00017,0.00022,0.00023,8e-04,0.00043,0.001,0.00118,0.00255,0.00128,0.00191,0.00074,0.00057,0.00015,0.00024,0.00021,0.00032,0.00023,0.00023,0.00017,0.00014,0.00017,0.00016,0.00021,0.00022,0.00062,0.00057,0.00044,0.00054,0.00097,0.00178,0.00182,0.00154,0.00121,0.00039)
spm <- get_spm_nechad(spm_rrs, sensor="occci")
plot(spm)

# example using a spatraster
library(terra)
data("example04_OCCCI.20240703.L3b.DAY.RRS.PANCAN")
spm_rrs <- terra::rast(example04_OCCCI.20240703.L3b.DAY.RRS.PANCAN$Rrs_665) # Daily Rrs_665 from OC-CCI multisensor product
spm <- get_spm_nechad(spm_rrs, sensor="occci")
plot(spm)

}
\references{
doi:10.1016/j.rse.2009.11.022
}
