% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPM.R
\name{get_spm_nechad_band}
\alias{get_spm_nechad_band}
\title{Get red waveband used in Nechad's SPM model}
\usage{
get_spm_nechad_band(sensor = "occci")
}
\arguments{
\item{sensor}{String, currently only occci is accepted}
}
\value{
Character vector in the form "Rrs_XXX" where XXX is the waveband in nanometers
}
\description{
Given a sensor name, get the waveband used in Nechad's SPM model (doi:10.1016/j.rse.2009.11.022)
}
