% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPM.R
\name{get_spm_nechad_coefs}
\alias{get_spm_nechad_coefs}
\title{Get Ap and Cp coefficients used in Nechad's SPM model}
\usage{
get_spm_nechad_coefs(sensor = "occci")
}
\arguments{
\item{sensor}{String, currently only occci is accepted}
}
\value{
Named list with Ap and Cp values
}
\description{
Given a sensor name, get the Ap and Cp coefficients used in Nechad's SPM model (doi:10.1016/j.rse.2009.11.022)
}
