% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_var}
\alias{get_var}
\title{Retrieve a variable}
\usage{
get_var(x, var = "chlor_a", nav = get_nav(x), form = c("matrix", "stars")[2])
}
\arguments{
\item{x}{ncdf4 object}

\item{var}{chr, the name of the variable to retrieve}

\item{nav}{list, a navigation list from \code{get_nav}}

\item{form}{chr, one of 'matrix' or 'stars'}
}
\description{
Retrieve a variable
}
