% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\name{gmt_palettes}
\alias{gmt_palettes}
\title{Colour Palettes}
\arguments{
\item{n}{Number of colours to return}
}
\value{
A vector of hexadecimal colours (#RRGGBB)
}
\description{
GMT Colour palettes
}
\details{
These are a few of the standard palettes included in GMT. See: https://docs.generic-mapping-tools.org/6.2/cookbook/cpts.html
}
\references{
GMT 6: Wessel, P., Luis, J. F., Uieda, L., Scharroo, R., Wobbe, F., Smith, W. H. F., & Tian, D. (2019).
The Generic Mapping Tools version 6. Geochemistry, Geophysics, Geosystems, 20, 5556–5564.
https://doi.org/10.1029/2019GC008515
}
