% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haversine.R
\name{haversine}
\alias{haversine}
\title{Haversine}
\usage{
haversine(lon1, lat1, lon2, lat2)
}
\arguments{
\item{lon1}{Decimal longitude of first point (numeric value).}

\item{lat1}{Decimal latitude of first point (numeric value).}

\item{lon2}{Decimal longitude of second point (numeric value).}

\item{lat2}{Decimal latitude of second point (numeric value).}
}
\value{
Distance between the points (in metres).
}
\description{
Calculate the haversine distance between 2 points in metres, given lat/lon for each (in decimal degrees).
}
