% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{lmp}
\alias{lmp}
\title{Get p-value}
\usage{
lmp(modelobject)
}
\arguments{
\item{modelobject}{Object of class lm}
}
\value{
F-test p-value (single numeric value)
}
\description{
Calculate the F-test p-value from a linear model object generated by the lm() function.
}
\references{
Source: Stephen Turner, Jan 2011
https://gettinggeneticsdone.blogspot.com/2011/01/rstats-function-for-extracting-f-test-p.html
}
