% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{make_raster_map}
\alias{make_raster_map}
\title{Plot a georeferenced raster on a map}
\usage{
make_raster_map(
  rast,
  title = NULL,
  xlim = c(-95, -42),
  ylim = c(39, 82),
  col_limits = NULL,
  cm = colorRampPalette(c("#00007F", "blue", "#007FFF", "cyan", "#7FFF7F", "yellow",
    "#FF7F00", "red", "#7F0000"))(100),
  na.value = "transparent",
  map_alpha = 0.8,
  nrow = 1
)
}
\arguments{
\item{rast}{Georeferenced RasterLayer or RasterStack}

\item{title}{Optional title of the map, for a single raster}

\item{xlim}{Longitude limits}

\item{ylim}{Latitude limits}

\item{col_limits}{Color scale limits}

\item{cm}{Color scale}

\item{na.value}{Color to use for NA values in raster}

\item{map_alpha}{Transparency of landmasses on map, 0-1}

\item{nrow}{Number of rows of plots, for raster stacks}
}
\value{
Raster or grid of rasters on maps with coastlines.
}
\description{
Plot a raster on a map. If a RasterStack is provided instead, facet_wrap will be used to plot the raster layers on separate maps, and nrow specifies the number of rows of maps (note: they will be plotted by row rather than by column). The names of the raster layers in the stack will be used as titles, and the maps will all use the same color scale.
}
\examples{
# use SGLI L2 data
library(raster)
library(sp)
# create a raster
data("example02_GC1SG1_202109031518L33309_L2SG_IWPRK_2000")
pts <- example02_GC1SG1_202109031518L33309_L2SG_IWPRK_2000
coordinates(pts) = ~lon+lat
tr <- rasterize(pts, raster(ext=extent(pts)), pts$chl, fun = mean, na.rm = TRUE)
# plot it on the map
make_raster_map(log10(tr),title=NULL)

}
