% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{monotonic_check}
\alias{monotonic_check}
\title{Check if a vector is monotonically increasing/decreasing}
\usage{
monotonic_check(v, direction = "increasing")
}
\arguments{
\item{v}{Numeric vector}

\item{direction}{String, either "increasing" or "decreasing"}
}
\value{
TRUE/FALSE
}
\description{
Check if a vector is monotonically increasing/decreasing
}
\examples{
v1 <- sample(1:100,size=10)
v2 <- 1:10
monotonic_check(v1)
monotonic_check(v2)
monotonic_check(v2, direction="decreasing")

}
