% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satfile_funcs.R
\name{nc_image_stats}
\alias{nc_image_stats}
\title{Get percent coverage of netCDF satellite image}
\usage{
nc_image_stats(
  file,
  w,
  e,
  s,
  n,
  var = "geophysical_data/Rrs_555",
  latvar = "navigation_data/latitude",
  lonvar = "navigation_data/longitude"
)
}
\arguments{
\item{file}{String, netCDF filename}

\item{w, e, s, n}{Numeric values, west/east/south/north in decimal degrees}

\item{var}{String, name of variable to check for percent coverage (full path to the variable within the netCDF, e.g. geophysical_variables/Rrs_555)}

\item{latvar}{String, full path and name of the latitude variable}

\item{lonvar}{String, full path and name of the longitude variable}
}
\value{
Dataframe containing filename, number of valid pixels, total number of pixels, and percent coverage. If file can't be read, "try-error" is returned instead.
}
\description{
Given a filename, W/E/S/N boundaries, and a variable name, check the percent coverage for that variable within the boundaries.
}
