% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCX.R
\name{oci}
\alias{oci}
\title{Hu Color Index Algorithm}
\usage{
oci(rrs, w)
}
\arguments{
\item{rrs}{Numeric matrix where rows = records, columns = 3 Rrs wavebands (blue, green, and red, in that order)}

\item{w}{Numeric vector, wavebands to use for the "blue", "green", and "red" bands, in that order (in nanometres), same as used in rrs}
}
\value{
Numeric vector of chla values
}
\description{
Given a matrix of Rrs (remote sensing reflectances), and the blue, green, and red wavebands to use, calculate chlorophyll using the Ocean Colour Index (OCI) from Hu et al (2012).
}
\details{
This is the algorithm used by NASA (as of Nov 2020) for chla concentrations < 0.15 mg /m^3, and blended with OCx for concentrations between 0.15 and 0.2.

The vectors of wavebands (w) for MODIS, SeaWiFS, and VIIRS-SNPP are c(443,555,667), c(443,555,670), and c(443,551,671), respectively.
}
\references{
Hu, Chuanmin & Lee, Zhongping & Franz, Bryan. (2012). Chlorophyll a algorithms for oligotrophic oceans: A novel approach based on three-band reflectance difference. Journal of Geophysical Research. 117. C01011. 10.1029/2011JC007395.
}
