% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHLA_OCX.R
\name{ocx_sse}
\alias{ocx_sse}
\title{OCx SSE}
\usage{
ocx_sse(params, insitu_chl, bandratio, alg_degree = 4, reg_method = 3)
}
\arguments{
\item{params}{Named numeric vector (names must be a0,a1,a2,a3,a4, or fewer if using a smaller alg_degree) - coefficients of the polynomial}

\item{insitu_chl}{Numeric vector of log10(in_situ_chla)}

\item{bandratio}{Numeric vector of log10(band_ratio), same length as insitu_chl}

\item{alg_degree}{Numeric value between 1 and 4 inclusive, polynomial degree of the algorithm (note that a 4th degree polynomial requires 5 coefficients, including the intercept term)}

\item{reg_method}{Number of the regression method used in lmodel2 (see ?lmodel2 for details, default = 3, Standard Major Axis)}
}
\value{
Numeric value, sum of squared error between in situ chl and satellite chl generated using the user-specified polynomial degree and coefficients
}
\description{
Calculate the sum of squared error between in situ chl and satellite OCx chl. Called from within optimize_ocx_coefs() with the optim() function to minimize the error and get the optimal set of coefficients for the set of in situ chl / band ratios.
}
