% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHLA_OCX.R
\name{ocxspmcor}
\alias{ocxspmcor}
\title{OCxSPM-Cor algorithm}
\usage{
ocxspmcor(rrs, spm, blues, green, coefs, use_443nm = FALSE)
}
\arguments{
\item{rrs}{Either: Numeric matrix where rows = records, columns = Rrs wavebands, with named columns ("Rrs_XXX", where XXX is a wavelength in nanometres), OR: RasterStack or RasterBrick or SpatRaster of rrs layers with stack layers following the same naming convention. Names must match c(blues, green), i.e. same names, from shortest waveband to longest.}

\item{spm}{Numeric vector of spm values, the same number of records as in rrs, OR: RasterLayer or SpatRaster of spm}

\item{blues}{Character vector of Rrs wavebands in the blue range (e.g. c("Rrs_443", Rrs_488")), matching column name(s) in rrs, maximum 3 options, arranged from shortest waveband to longest. Note that if use_443nm=FALSE, the 443nm waveband will be removed and another must be used in its place.}

\item{green}{String, Rrs waveband in the green range (e.g. "Rrs_547"), matching a column name in rrs}

\item{coefs}{Numeric vector of coefficients corresponding to terms in the polynomial (lowest degree to highest)}

\item{use_443nm}{Logical value, TRUE to make the 443nm band an option in the band ratio}
}
\value{
For matrix rrs: Numeric value (or vector) -- chlorophyll as computed by OCxSPM-Cor for the given Rrs and SPM, sensor, and coefficients. For RasterStack/Brick rrs: equivalent raster with OCxSPM-Cor chlorophyll-a.
}
\description{
Modified version of ocx() for use in waters with high SPM, tested in the Bay of Fundy. Input is the same as ocx(), with the requirement of an extra vector (or raster) of spm values.
}
\examples{

# example using a spatraster
library(terra)
data("example04_OCCCI.20240703.L3b.DAY.RRS.PANCAN")
# convert rrs to terra spatrasters
rrs <- lapply(example04_OCCCI.20240703.L3b.DAY.RRS.PANCAN, FUN=terra::rast) \%>\% unname() \%>\% do.call(what=c)
names(rrs) <- paste0("Rrs_",all_lambda$occci)
# calculate spm
spm_band <- get_spm_band(sensor="occci")
spm <- get_spm_nechad(rrs[[names(rrs)==spm_band]], sensor="occci")
# get the coefficients for ocxspmcor
coefs <- get_ocx_coefs("occci", region="bof", alg="ocxspmcor")
# get the wavebands used in the ocxspmcor algorithm for occci
lambdas <- get_ocx_bands("occci", use_443nm = FALSE)
# calculate ocxspmcor chl for each data point using those coefficients
sat_ocxspmcor_chl <- ocxspmcor(rrs=rrs, spm=spm, blues=lambdas$blues, green=lambdas$green, coefs=coefs)
plot(sat_ocxspmcor_chl)

}
\references{
doi:10.1007/s12237-024-01334-x
}
