% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCX.R
\name{optimize_ocx_coefs}
\alias{optimize_ocx_coefs}
\title{OCX coefficient optimization}
\usage{
optimize_ocx_coefs(
  data,
  alg_degree,
  params_guessed = c(a0 = 0.3, a1 = -3.8, a2 = -1, a3 = 1, a4 = 1),
  reg_method = 3
)
}
\arguments{
\item{data}{Dataframe containing 2 columns labelled x and y, where x = log10(in_situ_chla) and y = log10(band_ratio)}

\item{alg_degree}{Numeric value between 1 and 4 inclusive, polynomial degree of the algorithm (note that a 4th degree polynomial requires 5 coefficients, including the intercept term)}

\item{params_guessed}{Named numeric vector (names must be a0,a1,a2,a3,a4, or fewer if using a smaller alg_degree)}

\item{reg_method}{Number of the regression method used in lmodel2 (see ?lmodel2 for details, default = 3, Standard Major Axis)}
}
\value{
Numeric vector of optimized coefficients, for the term with the smaller degree to highest
}
\description{
Given a dataframe with log10(in_situ_chla) values in the first column and corresponding log10(satellite band_ratio) values in the second column, and the degree of the polynomial algorithm to use, find the best set of coefficients to fit the band ratio to the chla.
}
\details{
Note that this uses the full dataset for optimization (it's NOT separated into training/test/validation sets). See the example below for a method to use bootstrapping to get confidence intervals for the retrieved coefficients.
}
