% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{order_string}
\alias{order_string}
\title{Order the characters in a single string}
\usage{
order_string(x)
}
\arguments{
\item{r}{String}
}
\value{
String with individual characters in order
}
\description{
Order the characters in a single string
}
\examples{
x = "023521"
order_string(x)
}
