% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{pad0}
\alias{pad0}
\title{Pad a string or number with zeroes in front.}
\usage{
pad0(s, len)
}
\arguments{
\item{s}{A string or number.}

\item{len}{Numeric value, length of the final string.}
}
\value{
s with zeroes in front.
}
\description{
Given a number or string, pad the front with zeroes up to length len. This function is a simple wrapper for stringr::str_pad(s,width=len,side="left",pad="0").
}
\examples{
pad0(45,4)

}
