% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{pad_num}
\alias{pad_num}
\title{Number padding}
\usage{
pad_num(num, len)
}
\arguments{
\item{num}{The number.}

\item{len}{The length, single numeric value.}
}
\value{
The number as a string with leading zeroes.
}
\description{
Pad a number with leading zeroes to make it length "len".
}
