% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_funcs.R
\name{plot_pancan}
\alias{plot_pancan}
\title{Plot panCanadian L3b file}
\usage{
plot_pancan(
  vec,
  region = "pancan",
  ext = c(xmn = -147, xmx = -41, ymn = 39, ymx = 86),
  resolution = "4km",
  limits = NULL
)
}
\arguments{
\item{vec}{Numeric vector of data (must be in the same order as the the bins vector of the same resolution, whether pancan, nwa, nep, or gosl - see details).}

\item{region}{String, either "pancan", "nwa", "nep", or "gosl"}

\item{ext}{Named vector containing the boundaries of the resulting grid.}

\item{resolution}{String, either "4km" or "9km".}

\item{limits}{Limits of the colour scale (numeric vector, length 2).}
}
\value{
Raster containing variable values with coastlines.
}
\description{
Given a vector of data from a binned ocean colour file, plot it on a raster with coastlines.
}
\details{
This is used to get a quick look at a file from the panCanadian dataset, before manipulating the data. When the data is loaded, it's in vector format and in the same order as the bin vectors pancan_bins_4km, nwa_bins_4km, etc...
}
\examples{
# This is an example file with data in the same format as in a panCanadian dataset file.
data("example01_A2018252.L3b_DAY_CHL_POLY4_NWA")
dat <- example01_A2018252.L3b_DAY_CHL_POLY4_NWA
lon_lim <- lon_bounds[["NWA"]]
lat_lim <- lat_bounds[["NWA"]]
plot_pancan(log10(dat), region="nwa", ext=c(range(lon_lim),range(lat_lim)))

}
