% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satfile_funcs.R
\name{plot_swath}
\alias{plot_swath}
\title{Plot satellite swath boundary}
\usage{
plot_swath(
  lon,
  lat,
  color = "red",
  size = 2,
  mapfill = "darkgreen",
  mapcolor = "#444444",
  mapalpha = 0.5
)
}
\arguments{
\item{lon, lat}{2D numeric matrices, same size, containing the coordinates for the corresponding matrix containing the satellite data}

\item{color}{Color of the swath boundary}

\item{size}{Width of the swath outline}

\item{mapfill}{Color of the continents on the map}

\item{mapcolor}{Color of the outlines of countries on the map}

\item{mapalpha}{Transparency of the continents on the map}
}
\value{
Map with a polygon showing the satellite swath boundary
}
\description{
Plot satellite swath boundary
}
\examples{
# load the coordinate data from an example GCOM-C SGLI satellite file
data("example03_GC1SG1_201801031430Q31808_L2SG_IWPRK_3000")
str(longitude)
str(latitude)
plot_swath(lon=longitude, lat=latitude)
}
