% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{plus_minus}
\alias{plus_minus}
\title{Get vector surrounding number}
\usage{
plus_minus(x, n)
}
\arguments{
\item{x}{Integer, the number at the center of the vector}

\item{n}{Integer, the number of places to expand the vector on either side of x}
}
\value{
Numeric vector
}
\description{
Get a numeric vector of numbers from (x-n) to (x+n).
}
\examples{
plus_minus(4, 2)
}
