% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{raster_to_matrix}
\alias{raster_to_matrix}
\title{Convert raster stack to matrix}
\usage{
raster_to_matrix(r, rnames)
}
\arguments{
\item{r}{Raster stack with layer names}

\item{rnames}{Vector of names from the raster stack}
}
\value{
Numeric matrix with column names matching rnames
}
\description{
Given a raster stack with layer names, and a vector of names to extract from that raster stack, use the raster::getValues() function to extract the values from each layer and put each flattened layer in the column of a matrix.
}
