% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satfile_funcs.R
\name{read_h5_L3b}
\alias{read_h5_L3b}
\title{Read h5}
\usage{
read_h5_L3b(h5_file, var_name)
}
\arguments{
\item{h5_file}{String, h5 filename.}

\item{var_name}{String, name of variable to extract (example: "chlor_a").}
}
\value{
List containing the contents of the file, attributes, selected variable, bin list, and bin index.
}
\description{
Read h5 level-3 binned file contents.
}
