% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satfile_funcs.R
\name{read_h5_l3b_names}
\alias{read_h5_l3b_names}
\title{Get list of variable names, dimensions, and datatypes from an h5 file}
\usage{
read_h5_l3b_names(h5_file)
}
\arguments{
\item{h5_file}{String, h5 filename}
}
\value{
List containing variable names, their dimensions, and datatypes
}
\description{
Get list of variable names, dimensions, and datatypes from an h5 file
}
