% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pixEx.R
\name{read_pixEx}
\alias{read_pixEx}
\title{Read pixEx}
\usage{
read_pixEx(filename)
}
\arguments{
\item{filename}{String of .csv file location}
}
\value{
List containing the contents of the file and attributes
}
\description{
Read pixEx file output from SNAP Pixel extraction tool
}
