% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_funcs.R
\name{res_code_to_dist}
\alias{res_code_to_dist}
\title{Given L3b data resolution code, get distance}
\usage{
res_code_to_dist(resolution)
}
\arguments{
\item{resolution}{String representing the spatial resolution (see ?gen_nrows for list of accepted strings, e.g. "Q" for 250m or "4" for 4.6km)}
}
\value{
List with the distance in degrees and km for the selected resolution code
}
\description{
For a given spatial resolution code (see ?gen_nrows), get equivalent distance in km and degrees at the equator (i.e. assuming 1 degree = 111km).
}
\references{
https://oceancolor.gsfc.nasa.gov/docs/format/l3bins/
}
