% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{rm_dup}
\alias{rm_dup}
\title{Remove consecutive duplicate characters in a string}
\usage{
rm_dup(x, ch)
}
\arguments{
\item{x}{String}

\item{ch}{Duplicated character to remove}
}
\value{
String with duplicates removed.
}
\description{
Remove consecutive duplicate characters in a string
}
\examples{
rm_dup("this_is_my__string", "_")
rm_dup("this.is.my..string", ".")

}
