% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{rmse}
\alias{rmse}
\title{Calculate RMSE}
\usage{
rmse(x, y)
}
\arguments{
\item{x}{Numeric vector of data.}

\item{y}{Numeric vector of data to compare to x, same length.}
}
\value{
RMSE, single numeric value.
}
\description{
Compute root mean square error (RMSE).
}
