% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHLA_GSM.R
\name{rrs_above_to_below_sea_level}
\alias{rrs_above_to_below_sea_level}
\title{Convert Rrs below sea level}
\usage{
rrs_above_to_below_sea_level(rrs)
}
\arguments{
\item{rrs}{Numeric value, vector, or matrix of Rrs}
}
\value{
Numeric value, vector, or matrix of corresponding Rrs below sea level
}
\description{
Given a set of remote-sensing reflectances (Rrs) above sea level, convert them to equivalent values just below the surface.
}
