% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{set_limits}
\alias{set_limits}
\title{Set values outside range to upper/lower limits}
\usage{
set_limits(v, lower = -Inf, upper = Inf, na.rm = TRUE)
}
\arguments{
\item{v}{Numeric vector}

\item{lower}{Lower limit, single numeric value}

\item{upper}{Upper limit, single numeric value}

\item{na.rm}{How to treat NA values}
}
\value{
Numeric vector where values beyond the accepted range have been converted to the lower/upper limits
}
\description{
Given a numeric vector and lower and upper limits of allowed values, set the values outside the accepted range to the nearest limit.
}
\details{
Other methods to get the same result:
\enumerate{
\item \code{ifelse(v>upper,upper,ifelse(v<lower,lower,v))}
\item \code{v[v<upper] <- upper;}
\code{v[v>lower] <- lower}
}
}
\examples{
v <- 1:10
l <- 3
v <- set_limits(v,l)

v <- rnorm(1e5)
l <- -2
u <- 2

ptm <- Sys.time()
v <- set_limits(v,l,u)
print(Sys.time()-ptm)

ptm <- Sys.time()
v[v<l] <- l
v[v>u] <- u
print(Sys.time()-ptm)

}
