% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_fn.R
\name{shift_line}
\alias{shift_line}
\title{Shift line}
\usage{
shift_line(x, y, dist = 1, dir = "up")
}
\arguments{
\item{x}{Numeric vector, length >=2.}

\item{y}{Numeric vector, same length as x.}

\item{dist}{Numeric value, how far to shift the line.}

\item{dir}{String, either "up" or "down".}
}
\value{
Named list containing the new x and y vectors.
}
\description{
Given an x vector of at least 2 points, a corresponding y vector of equal length, and a distance to shift the line, shift a SLANTED line "up" or "down", in the direction perpendicular to the line.
}
