% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{sinh_trans}
\alias{sinh_trans}
\title{sinh ggplot scale transformation}
\usage{
sinh_trans()
}
\description{
This can be used to do an sinh transformation of a ggplot scale. To do an asinh transformation, use "asinh" instead of "sinh" (see ?asinh_trans).
}
\examples{
library(ggplot2)
ggplot(data.frame(x=1:10,y=1:10), aes(x=x,y=y)) +
    geom_point() +
    scale_y_continuous(trans="sinh")

}
