% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_gaps.R
\name{sparkle_fill}
\alias{sparkle_fill}
\title{Sparkle fill: fill small holes in raster}
\usage{
sparkle_fill(x, min_sides, fun, ...)
}
\arguments{
\item{x}{RasterLayer to fill gaps for}

\item{min_sides}{Number of non NA sides required (e.g. 5 out of a possible 8 side pixels in box around "sparkle")}

\item{fun}{String of function to fill gap. Current options are "mean" and "median"}
}
\value{
Filled RasterLayer
}
\description{
Given a RasterLayer with small gaps (sparkles), fill using specs defined in "min_sides" and "fun"  Gaussian.
}
