% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_funcs.R
\name{var_to_rast}
\alias{var_to_rast}
\title{Generate raster of L3b data}
\usage{
var_to_rast(
  df,
  resolution = "4km",
  ext = c(xmn = -147, xmx = -41, ymn = 39, ymx = 86)
)
}
\arguments{
\item{df}{Dataframe with 2 columns: bin numbers, and variable values (note: the dataframe does not have to be in order of bin number).}

\item{resolution}{String, either "4km" or "9km".}

\item{ext}{Named vector containing the boundaries of the resulting grid.}
}
\value{
Global raster containing variable values.
}
\description{
Create 2d grid of a variable, given a dataframe with 2 columns, one for bin numbers and one for variable values.
}
