% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{vector_stats}
\alias{vector_stats}
\title{Calculate statistics}
\usage{
vector_stats(x)
}
\arguments{
\item{x}{Numeric vector of data.}
}
\value{
Named numeric vector of statistics, length 5.
}
\description{
Calculate the statistics on a vector: Number of valid observations, mean, median, standard deviation, and coefficient of variation.
}
