% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_funcs.R
\name{week8}
\alias{week8}
\title{Get the 8-day week number of a given date or date vector}
\usage{
week8(dateval, ..)
}
\arguments{
\item{dateval}{String or formatted date, e.g. as.Date("2020-01-04")}
}
\value{
numeric value of week number
}
\description{
This converts dates to an 8-day week number as used by NASA. Use like other lubridate functions like lubridate::week().
}
