% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_funcs.R
\name{week8_date}
\alias{week8_date}
\title{Date from 8-day week number}
\usage{
week8_date(yearnum, weeknum)
}
\arguments{
\item{yearnum}{Numeric year number or vector}

\item{weeknum}{Numeric week number or vector}
}
\value{
date value formatted with as.Date()
}
\description{
This converts a week number (8 days per week) to its corresponding date. Year is required as this will vary for leap years. 8 days is the standard number of days in a composite "week" time period as used for ocean colour by NASA.
}
\examples{
week8_2md(2010, 8)
week8_2md(yearnum = c(2010,2010), weeknum = c(8,9))
}
