% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{add_matrices}
\alias{add_matrices}
\title{Matrix addition}
\usage{
add_matrices(..., na.rm = TRUE)
}
\arguments{
\item{...}{Numeric matrices to add together (they must be the same size)}
}
\value{
The numeric matrix that is the sum of the input matrices (same shape)
}
\description{
Add matrices together, specifying na.rm=TRUE or FALSE for each corresponding cell.
}
