% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_funcs.R
\name{binlatlon}
\alias{binlatlon}
\title{Get bin, longitude, latitude}
\usage{
binlatlon(resolution = "4km", lonlim = c(-180, 180), latlim = c(-90, 90))
}
\arguments{
\item{resolution}{Spatial resolution for binned grid (either 1km, 4km, 9km, or 111km)}

\item{lonlim}{Minimum and maximum longitude of the area of interest}

\item{latlim}{Minimum and maximum latitude of the area of interest}
}
\value{
Dataframe with 3 columns: bin, longitude, latitude
}
\description{
This creates a dataframe containing the bin number, longitude, and latitudes for the full globe, for a given resolution (4.64km, 9.28km, or 111km), using the Integerized Sinusoidal Binning Scheme used by NASA OBPG for their level-3 binned satellite files (e.g. MODIS-Aqua). More info here: https://oceancolor.gsfc.nasa.gov/docs/format/l3bins/
}
