% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_funcs.R
\name{collapse_window}
\alias{collapse_window}
\title{Get single value from pixel window}
\usage{
collapse_window(
  var,
  input_ws,
  output_ws,
  min_pix,
  useCVfilter = TRUE,
  max_CV = 0.15,
  stat_to_use = "median"
)
}
\arguments{
\item{var}{Numeric vector or square matrix, length input_ws^2}

\item{input_ws}{Integer, length of one side of the input square pixel box}

\item{output_ws}{Integer, length of one side of the output square pixel box (<= input_ws)}

\item{min_pix}{Integer, minimum number of valid pixels required}

\item{useCVfilter}{TRUE/FALSE - if median CV (coefficient of variation, filteredSD/filteredMean) is > max_CV, return NA?}

\item{max_CV}{Maximum allowed CV}

\item{stat_to_use}{String, either median, mean, or geomean - the statistic used on the final filtered set of var}
}
\value{
Single numeric value representing the input window
}
\description{
Given a square matrix of pixel values, return a single value representing the whole box.
}
\details{
This is used when reading L2 satellite points matched to in situ data in order to get a single pixel value representing the box (e.g. 5x5 pixels) around the center pixel closest to the in situ sampling location.
}
