% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_funcs.R
\name{deg_to_dm}
\alias{deg_to_dm}
\title{Convert a latitude or longitude from decimal degrees to degrees, decimal minutes}
\usage{
deg_to_dm(decdeg)
}
\arguments{
\item{decdeg}{number in decimal degrees}
}
\value{
Returns a data.frame() with columns indicating degrees and decimal minutes.
}
\description{
To reverse this, use \code{dms_to_deg()} and leave the "sec" argument blank.
}
\details{
For degrees, minutes, seconds use \code{deg_to_dms()}.
}
\examples{
deg_to_dm(-52.80833)

}
