% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satfile_funcs.R
\name{download_oc}
\alias{download_oc}
\title{Download ocean colour files}
\usage{
download_oc(filenames, urls)
}
\arguments{
\item{filenames}{Filename to download}

\item{urls}{URLs of files corresponding to filenames}
}
\value{
Data.frame with metadata and URLs of files to download.
}
\description{
Download file lists from functions get_imglist_multi() and get_imglist_l1()
Uses wget, and you need to have the OBPG .netrc credentials - see "Download Methods" here: https://oceancolor.gsfc.nasa.gov/data/download_methods/
}
\examples{
filename = "A2003001162000.L2_LAC_OC.nc"
url = "https://oceandata.sci.gsfc.nasa.gov/cmr/getfile/A2003001162000.L2_LAC_OC.nc"
download_oc(filename, url)
}
