% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_to_raster.R
\name{gen_bin_grid}
\alias{gen_bin_grid}
\title{Generate bin grid}
\usage{
gen_bin_grid(start_num)
}
\arguments{
\item{start_num}{Vector of numbers for the bins at the start of each row, generated by gen_start_bin().}
}
\value{
Global raster containing bin numbers.
}
\description{
This generates a raster of bin numbers based on the bin numbers at the start of each row.
Note that the latitudes in the array go from +90 to -90, while the start_num vector goes from -90 to 90.
}
