% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_funcs.R
\name{gen_bin_grid}
\alias{gen_bin_grid}
\title{Generate bin grid}
\usage{
gen_bin_grid(
  resolution = "4km",
  ext = c(xmn = -147, xmx = -41, ymn = 39, ymx = 86)
)
}
\arguments{
\item{resolution}{String, either "1km", "4km", "9km", or "111km".}

\item{ext}{Named vector containing the boundaries of the resulting grid (xmn, xmx, ymn, ymx).}
}
\value{
Global raster containing bin numbers.
}
\description{
This generates a raster of bin numbers for the selected extent and resolution.
}
