% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_funcs.R
\name{gen_bin_grid}
\alias{gen_bin_grid}
\title{Generate bin grid}
\usage{
gen_bin_grid(
  resolution = "4",
  ext = c(xmn = -147, xmx = -41, ymn = 39, ymx = 86),
  rast = TRUE
)
}
\arguments{
\item{resolution}{String indicating spatial resolution, see ?gen_nrows for list of accepted strings.}

\item{ext}{Named vector containing the boundaries of the resulting grid (xmn, xmx, ymn, ymx).}

\item{rast}{TRUE/FALSE, should the resulting bin matrix be converted to raster?}
}
\value{
Global raster containing bin numbers.
}
\description{
This generates a matrix or raster of bin numbers for the selected extent and resolution.
}
