% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_to_raster.R
\name{gen_start_bin}
\alias{gen_start_bin}
\title{Generate start bin vector}
\usage{
gen_start_bin(nrows = 4320)
}
\arguments{
\item{nrows}{Number of rows in the resulting global raster of bin numbers.}
}
\value{
Numeric vector of starting bins.
}
\description{
This generates a vector of bin numbers at the start of each row on a global grid, given the number of rows.
Example: For 4km resolution, 4320 rows; for 9km resolution, 2160 rows
}
