% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_fn.R
\name{geoMean}
\alias{geoMean}
\title{Geometric mean}
\usage{
geoMean(x, ...)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Geometric mean value of the input data
}
\description{
As opposed to arithmetic mean, this calculates the mean of the log transformed data and converts it back. Use with lognormal distributions like chlorophyll-a
}
