% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_funcs.R
\name{geoSD}
\alias{geoSD}
\title{Geometric standard deviation}
\usage{
geoSD(x, ...)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Geometric mean value of the input data
}
\description{
This calculates the geometric SD
}
