% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHLA_OCX.R
\name{get_ocx_coefs}
\alias{get_ocx_coefs}
\title{Get predefined coefficients for OCx}
\usage{
get_ocx_coefs(sensor, region = "global", alg = "ocx")
}
\arguments{
\item{sensor}{String, either "modis", "seawifs", or "viirs" (note: "modis" is MODIS-Aqua, and "viirs" is VIIRS-SNPP)}

\item{region}{String, either "global" (for ocx), or "nwa" or "nep" (for ocx or poly1 to poly4)}

\item{alg}{String, either "poly1", "poly2", "poly3", "poly4", or "ocx"}
}
\value{
Numeric vector of polynomial coefficients, from the coefficient on the lowest degree term to highest degree.
}
\description{
Library of existing optimized coefficients for the polynomial band ratio algorithm.
"ocx" returns the coefficients of the global ocean colour algorithm used by NASA (as of Nov 2020) for the selected sensor. "poly1" to "poly4" return the regionally-optimized coefficients for polynomial algorithms of degrees 1 to 4 for the selected sensor, for the two available regions: nwa (Northwest Atlantic) or nep (Northeast Pacific).
}
\references{
Clay, S.; Peña, A.; DeTracey, B.; Devred, E. Evaluation of Satellite-Based Algorithms to Retrieve Chlorophyll-a Concentration in the Canadian Atlantic and Pacific Oceans. Remote Sens. 2019, 11, 2609.
https://www.mdpi.com/2072-4292/11/22/2609
}
