% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_funcs.R
\name{get_season}
\alias{get_season}
\title{Get season name}
\usage{
get_season(date)
}
\value{
String (either Spring, Summer, Fall, or Winter)
}
\description{
Given a date object, get the name of the season.
}
\details{
Feb-Apr = Spring,
May-Jul = Summer,
Aug-Oct = Fall,
Nov-Feb = Winter,
}
