% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{make_raster_map}
\alias{make_raster_map}
\title{Plot a georeferenced raster on a map}
\usage{
make_raster_map(
  rast,
  title = NULL,
  xlim = c(-95, -42),
  ylim = c(39, 82),
  col_limits = NULL,
  cm = colorRampPalette(c("#00007F", "blue", "#007FFF", "cyan", "#7FFF7F", "yellow",
    "#FF7F00", "red", "#7F0000"))(100)
)
}
\arguments{
\item{rast}{Georeferenced raster}

\item{title}{Optional title of the map}

\item{xlim}{Longitude limits}

\item{ylim}{Latitude limits}

\item{col_limits}{Color scale limits}

\item{cm}{Color scale}
}
\value{
Raster containing variable values with coastlines.
}
\description{
Plot a georeferenced raster on a map
}
\examples{
# use SGLI L2 data
library(raster)
library(sp)
# create a raster
data("example02_GC1SG1_202109031518L33309_L2SG_IWPRK_2000")
pts <- example02_GC1SG1_202109031518L33309_L2SG_IWPRK_2000
coordinates(pts) = ~lon+lat
tr <- rasterize(pts, raster(ext=extent(pts)), pts$chl, fun = mean, na.rm = TRUE)
# plot it on the map
make_raster_map(log10(tr),title=NULL)

}
