% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCX.R
\name{ocx}
\alias{ocx}
\title{OCX algorithm}
\usage{
ocx(rrs, blues, green, coefs, use_443nm)
}
\arguments{
\item{rrs}{Numeric matrix where rows = records, columns = Rrs wavebands, with named columns ("Rrs_XXX", where XXX is a wavelength in nanometres)}

\item{blues}{Character vector of Rrs wavebands in the blue range (e.g. "Rrs_488"), matching column name(s) in rrs, maximum 3 options}

\item{green}{String, Rrs waveband in the green range (e.g. "Rrs_547"), matching a column name in rrs}

\item{coefs}{Numeric vector of coefficients corresponding to terms in the polynomial (lowest degree to highest)}

\item{use_443nm}{Logical value, TRUE to make the 443nm band an option in the band ratio}
}
\value{
Numeric value (or vector), chlorophyll as computed by OCX for the given Rrs, sensor, and coefficients.
}
\description{
Given a set of coefficients, calculate chlorophyll using a polynomial band ratio algorithm. See ?optimize_ocx_coefs for example.
}
\references{
Original paper:

O'Reilly, John & Maritorena, S. & Mitchell, B.G. & Siegel, David & Carder, Kendall & Garver, S.A. & Kahru, Mati & Mcclain, Charles. (1998). Ocean color chlorophyll algorithms for SeaWiFS. Journal of Geophysical Research. 103. 937-953.

Reference for regional OCX algorithms tuned to Atlantic and Pacific Canadian coasts:

Clay, S.; Peña, A.; DeTracey, B.; Devred, E. Evaluation of Satellite-Based Algorithms to Retrieve Chlorophyll-a Concentration in the Canadian Atlantic and Pacific Oceans. Remote Sens. 2019, 11, 2609.
https://www.mdpi.com/2072-4292/11/22/2609
}
