% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_funcs.R
\name{plot_rast_from_bin}
\alias{plot_rast_from_bin}
\title{Plot panCanadian L3b file}
\usage{
plot_rast_from_bin(
  vec,
  ext = c(xmn = -147, xmx = -41, ymn = 39, ymx = 86),
  resolution = "4km",
  limits = c(-Inf, Inf)
)
}
\arguments{
\item{vec}{Numeric vector of data.}

\item{ext}{Named vector containing the boundaries of the resulting grid.}

\item{resolution}{String, either "4km" or "9km".}

\item{limits}{Limits of the colour scale (numeric vector, length 2).}
}
\value{
Raster containing variable values with coastlines.
}
\description{
Given a vector of data from a binned panCanadian ocean colour file, plot it on a raster with coastlines.
}
