% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{rmse}
\alias{rmse}
\title{Calculate RMSE}
\usage{
rmse(x, y, dolog10 = FALSE)
}
\arguments{
\item{x}{Numeric vector of data.}

\item{y}{Numeric vector of data to compare to x, same length.}

\item{dolog10}{TRUE/FALSE, should x and y be logged (base 10)? (i.e. calculate RMSLE?)}
}
\value{
Single numeric value.
}
\description{
Compute root mean square error (RMSE).
}
