% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{same_scales}
\alias{same_scales}
\title{Make scales of patchwork object the same}
\usage{
same_scales(
  p,
  xsame = TRUE,
  ysame = TRUE,
  xmin = -Inf,
  xmax = Inf,
  ymin = -Inf,
  ymax = Inf
)
}
\arguments{
\item{p}{Patchwork object}

\item{xsame}{TRUE/FALSE, should x axes be the same?}

\item{ysame}{TRUE/FALSE, should y axes be the same?}

\item{xmin}{Optional minimum value for x axis (see details)}

\item{xmax}{Optional maximum value for x axis (see details)}

\item{ymin}{Optional minimum value for y axis (see details)}

\item{ymax}{Optional maximum value for y axis (see details)}
}
\value{
Patchwork object, with axes adjusted
}
\description{
Given a patchwork object (see patchwork library), make the x and/or y scales of the plots the same.
}
\details{
Note that if you want to manually adjust scales (ymin, ymax, etc), you should do it in each plot using, for example, scale_y_continuous limits. Here, xmin/xmax are only used if xsame=TRUE, and ymin/ymax are only used if ysame=TRUE.
}
