% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_fn.R
\name{shifted_gaussian}
\alias{shifted_gaussian}
\title{Shifted Gaussian Curve}
\usage{
shifted_gaussian(tv, B0, beta = 0, h, sigma, tmax)
}
\arguments{
\item{tv}{Numeric vector}
}
\value{
Numeric vector containing the points along the Gaussian.
}
\description{
Given a vector tv, and parameters of a Gaussian curve, calculate a shifted Gaussian.
}
