% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_to_raster.R
\name{var_to_rast}
\alias{var_to_rast}
\title{Generate variable grid}
\usage{
var_to_rast(df, resolution = "4km")
}
\arguments{
\item{df}{Dataframe with 2 columns: bin numbers, and variable values.}

\item{resolution}{String, either "4km" or "9km".}
}
\value{
Global raster containing variable values.
}
\description{
Create 2d grid of a variable, given a dataframe with 2 columns, one for bin numbers and one for variable values.
}
