% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_funcs.R
\name{var_to_rast}
\alias{var_to_rast}
\title{Generate raster of L3b data}
\usage{
var_to_rast(
  df,
  resolution = "4km",
  ext = c(xmn = -147, xmx = -41, ymn = 39, ymx = 86)
)
}
\arguments{
\item{df}{Dataframe with 2 columns: first column="bin" and second column is the variable name (note: the dataframe does not have to be sorted in order of bin number).}

\item{resolution}{String, either "4km" or "9km".}

\item{ext}{Named vector containing the boundaries of the resulting grid (xmn, xmx, ymn, ymx).}
}
\value{
Raster containing the variable data.
}
\description{
Given a dataframe with 2 columns (bin number and variable), create the corresponding raster for visualization.
}
\details{
As of 2022-10-12, this function combines the code from gen_start_bin() and gen_bin_grid() in order to speed it up by subsetting it by latitude.
}
\examples{
# make a bathymetry raster for the Northwest Atlantic
var_to_rast(df = get_bins(region="nwa",variables=c("bin","bathymetry")),
            ext = c(lon_bounds$NWA, lat_bounds$NWA))
}
\references{
See https://oceancolor.gsfc.nasa.gov/docs/format/l3bins/ for more information on bin numbers.
}
