% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_funcs.R
\name{week8}
\alias{week8}
\title{8-day week number}
\usage{
week8(dateval)
}
\arguments{
\item{dateval}{String or date formatted with as.Date()}
}
\value{
numeric value of week number
}
\description{
This converts a date or date vector to its corresponding 8-day week number(s). 8 days is the standard number of days in a composite "week" time period as used for ocean colour by NASA. Use similarly to lubridate functions like lubridate::week().
}
\examples{
week8("2020-03-27")
week8(your_data_frame$Dates)
}
